Add tabbed UI layout and fix display size
