+++ MainWindow.xaml
@@ -1,28 +1,22 @@
-﻿<Window x:Class="MainWindow"
+<Window x:Class="MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         mc:Ignorable="d"
-    xmlns:controls="clr-namespace:Controls"
-        Title="APIテストツール" Height="450" Width="800">
-    <Grid Margin="10">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="2*"/>
-        </Grid.RowDefinitions>
-        <controls:UrlInput x:Name="UrlInputControl" Grid.Row="0" Margin="0,0,0,5"/>
-        <StackPanel Orientation="Horizontal" Grid.Row="1" Margin="0,0,0,5">
-            <controls:MethodSelecter x:Name="MethodSelecterControl" Width="100"/>
-            <controls:AuthInput x:Name="AuthInputControl" Margin="10,0,0,0"/>
-        </StackPanel>
-        <controls:HeaderInput x:Name="HeaderInputControl" Grid.Row="2" Margin="0,0,0,5"/>
-        <controls:BodyInput x:Name="BodyInputControl" Grid.Row="3" Margin="0,0,0,5"/>
-        <Button x:Name="SendButton" Grid.Row="4" Content="送信" Width="100" HorizontalAlignment="Right" Click="SendButton_Click"/>
-        <controls:ResponseOutput x:Name="ResponseOutputControl" Grid.Row="5"/>
-    </Grid>
+        xmlns:controls="clr-namespace:Controls"
+        Title="APIテストツール" Height="600" Width="800">
+    <DockPanel>
+        <Button x:Name="AddTabButton" Content="タブ追加" DockPanel.Dock="Top" Margin="5" Click="AddTabButton_Click"/>
+        <TabControl x:Name="MainTabControl">
+            <TabItem Header="TestAll">
+                <StackPanel Margin="10">
+                    <TextBlock Text="全体実行機能は未実装です。"/>
+                </StackPanel>
+            </TabItem>
+            <TabItem Header="Request1">
+                <controls:ApiRequestTab/>
+            </TabItem>
+        </TabControl>
+    </DockPanel>
 </Window>
